use --dart-define-from-file

+++ Makefile
@@ -3,12 +3,12 @@ clean:
 	flutter pub get
 
 install-ios-release: clean
+	test -f .env.json || (echo "Missing .env.json. Copy env.example.json to .env.json and fill values." && exit 1)
 	flutter build ipa --export-method development \
-		--dart-define=ADMOB_APP_ID=ca-app-pub-1679051359335856~1094547017 \
-		--dart-define=ADMOB_BANNER_IOS=ca-app-pub-1679051359335856/4194550743
+		--dart-define-from-file=.env.json
 	flutter install --use-application-binary=build/ios/ipa/SwingGroove\ Golf.ipa
 
 build-ipa: clean
+	test -f .env.json || (echo "Missing .env.json. Copy env.example.json to .env.json and fill values." && exit 1)
 	flutter build ipa --export-method app-store \
-		--dart-define=ADMOB_APP_ID=ca-app-pub-1679051359335856~1094547017 \
-		--dart-define=ADMOB_BANNER_IOS=ca-app-pub-1679051359335856/4194550743
+		--dart-define-from-file=.env.json
